Improve redis configuration for openedx and richie
Improve redis and redis sentinel configuration for the 2 stacks:
richie and openedx.
Allow any size of redis sentinel replicas.
We don't need to configure specific location for redis sentinel.
Openedx can use the HAProxy load balancer to use correct redis backend.
The HAProxy will route the traffic to the correct redis master instance
that sais that it's the master and where >50% of replicas sentinel agree
who redis instance is the master.