Reconfigure docker containers memory
Configure a soft (reservation) and hard limit memory limits
for docker containers on all stacks.
Docker can use the soft limit if it starts to have too low resources,
that is why we configure the soft limit of each container to
a value that should be enough for the service to still run.
GN-1011