Improve openedx nginx ssl w/ http2 & wildcard ssl
Allow to configure wildcard ssl domain only on http 1.1.
Allow to configure ssl with http2.
Add new nginx port for https and http2 on openedx.
Nginx will activate http2 for all pairs of <IP>:<port> site.
That is why we had to add a new port for https and http2.
So we have this nginx ports:
- 80 http
- 443 https without http2
- 8443 https with http2
GN-1169