feat: add a new VIP for ClickHouse Cluster
For performance reasons and to be similar to other Database cluster, it was decided to use a VIP also for ClickHouse Cluster.
This VIP will balance between all Cluster nodes.
This won't introduce latency between apps and the database itself.
This will replace the HAProxy-Clickhouse deployment that was deployed inside the Kubernetes.

fccn/nau-technical#705